Add files via upload
This is a program that takes a given number of particles and evolves time forward for a set number of collisions. What's returned in the output is the time and position of each particle at each collision. 

WARNING: the code becomes quite buggy for high velocities confined in small spaces but works very well in a reasonable regime of v within 100 and the pre-defined width of 20 although more conventional choices for velocity are of the order of 1 -> 10.